+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Fixed the order of items in ``WikiList.fullitems``. (#72)
 - Fixed and improved a few edge cases in ``Table.caption``. (pr #81)
 - Fixed handling of external links within definition lists. (pr #83)